requirements.txt
pbr>=0.6,!=0.7,<1.0
argparse
icalendar
jinja2
pyyaml
python-dateutil
